Misc stream fixes and failing fixes. Vanilla test is now added as a unit test.
Halt - now throws exception (PrologHalt), making it testable.
Char conversion now throws correct error
Abort honors 'close on abort' (currently not tested)
Stub created for seeking to end of stream (will change in future)
Streams now has capability of changing defaults - needed so we can change the default stream for the vanilla test. This complicated the logic of prohibiting the closing of a default stream.
Added capability of grepping for a string marker in failed tests.